LEMUR analysis for neighborhoods on predicted gene differential expression with conformal methods

Runs a LEMUR analysis, identifies differential-expression neighborhoods, and applies conformal procedures to a set of genes of interest. The data are split into training, calibration, and test sets based on the covariates specified in the LEMUR design formula.

conformeR(
  sce,
  design_lemur = ~1,
  contrast_column,
  contrast_levels = NULL,
  n_embedding = 60,
  use_assay = "logcounts",
  test_fraction_lemur = 0.2,
  cp_train_frac = 0.5,
  cp_cal_frac = 0.25,
  cp_alpha = 0.05,
  genes_of_interest,
  what = c("conf_selection", "conf_clustering"),
  n_cores = 1,
  verbose = TRUE
)

Arguments

sce

A SingleCellExperiment object containing the single-cell expression data.

design_lemur

A formula specifying the design used by LEMUR.

contrast_column

A column name specifying the comparison used for differential-expression testing in LEMUR.

n_embedding

Number of dimensions in the LEMUR embedding.

use_assay

Character string specifying the assay to use.

test_fraction_lemur

Fraction of cells used as the LEMUR test set.

cp_train_frac

Fraction of cells used for conformal training.

cp_cal_frac

Fraction of cells used for conformal calibration.

cp_alpha

Miscoverage level for the conformal procedures.

genes_of_interest

Character vector containing the names of genes to include in the conformal analysis.They need to correspond to the row names of the sce.

what

Character vector specifying the conformal procedure(s) to run.

n_cores

Number of cores used for parallel processing.

verbose

Logical indicating whether to display progress information.

constrast_levels

when contrast_column has more than two levels, specify which two are tested again each other.

eps_corruption

Label corruption level estimate used in the conformal procedures (vector of length length(genes_of_interest)).

Value

A list containing:

fit_lemur

The LEMUR test-set object restricted to the genes of interest.

nei_lemur

The LEMUR neighborhoods for the genes of interest in the test set.

conf_results

The results of the conformal procedures for the genes of interest.

Details

The conformal procedures are applied to a subset of genes of interest. Genes are processed in chunks to allow parallel computation.